// app.client/api/models/Department.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Department = {
    departmentID: string;
    name: string;
    description?: string;
    email?: string;
    createdBy?: string;
    createdDate?: string;
};