// app.client/api/models/UpdateDepartmentCommand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CreateDepartmentCommand } from './CreateDepartmentCommand';
export type UpdateDepartmentCommand = (CreateDepartmentCommand & {
    departmentID: string;
});